add/campains controller

+++ controllers/contact.js
@@ -1,6 +1,6 @@
 const asyncHandler = require("express-async-handler");
-const Product = require("../model/Product");
 const Contact = require("../model/Contact");
+const User = require("../model/User");
 
 //İletişim mesajlarını getirme
 exports.getContacts = asyncHandler(async (req, res) => {
@@ -14,6 +14,10 @@ exports.getContacts = asyncHandler(async (req, res) => {
 
 //Mesaj oluşturma
 exports.createContact = asyncHandler(async (req, res) => {
+  const userFound = await User.findById(req.userAuth._id);
+  if (!userFound) {
+    throw new Error("Kullanıcı bulunamadı");
+  }
   //* Create contact
   const contact = await Contact.create(req.body);
   //send the response
@@ -26,6 +30,10 @@ exports.createContact = asyncHandler(async (req, res) => {
 
 //mesaj silme
 exports.deleteContact = asyncHandler(async (req, res) => {
+  const userFound = await User.findById(req.userAuth._id);
+  if (!userFound) {
+    throw new Error("Kullanıcı bulunamadı");
+  }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(201).json({
     status: "Başarılı",
